requirements.txt
absl-py==2.0.0
astunparse==1.6.3
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
filelock==3.12.0
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.39.3
gast==0.5.4
google-auth==2.23.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.9.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.14.0
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.24.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybase64==1.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
sympy==1.11.1
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
threadpoolctl==3.2.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.5
Werkzeug==2.3.3
wrapt==1.14.1
